feat(filters): Get hydrated filtering doc
